+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.0.7
+control over stream
+control over md5
+return blob as Blob for web and as PlatformFile for non-web.
+
 ## 0.0.6
 cancel support.